Former-commit-id: 2f7f8466fedb4b3ea9507d62a8f1f43966f887d0

+++ Projects/UnitTests/Classes/OldTests/StaticWebViewTest.cpp
@@ -91,7 +91,7 @@ void StaticWebViewTest::LoadResources()
     DVASSERT(font);
 	font->SetSize(20);
 
-	const int32 w = 40;
+	const float32 w = 40;
 
     CreateUIButton(finishTestButton, font, Rect(0, 510, 300, w),
         L"Finish Test", &StaticWebViewTest::OnButtonPressed);
